refactored and simplified
