+++ frontlook_csharp_library/FL_Odbc_Helper/FL_odbc_helper.cs
@@ -4,9 +4,9 @@ using System.Data.OleDb;
 using System.Diagnostics.CodeAnalysis;
 using System.Windows;
 
-namespace frontlook_csharp_library.Data_Manager.FL_database_helper
+namespace frontlook_csharp_library.FL_Odbc_Helper
 {
-    public class FL_Odbc_Helper
+    public class FL_Odbc_Manager
     {
         public static DataTable FL_get_odbc_datatable(string constring, string query)
         {
