+++ frontlook_csharp_library/FL_general/FL_string_helper/FL_String_Validator.cs
@@ -2,7 +2,7 @@
 
 // ReSharper disable Checkstring namespace
 
-namespace frontlook_csharp_library.FL_general.FL_string_helper
+namespace frontlook_csharp_library.FL_General.FL_string_helper
 {
     /// <summary>
     /// String Maintainer
